[ICU-3500] Fix some compiler warnings X-Commit-URL: https://ssl.icu-project.org/trac/changeset/15886 Former-commit-id: c8711529a1e32a58f7a9c6d7950cdcec95eb8fa3

+++ icu4c/source/common/locid.cpp
@@ -1148,7 +1148,7 @@ Locale::getLocaleCache(void)
     umtx_unlock(NULL);
 
     if (needInit) {
-        Locale *tLocaleCache = new Locale[eMAX_LOCALES];
+        Locale *tLocaleCache = new Locale[(int)eMAX_LOCALES];
         if (tLocaleCache == NULL) {
             return NULL;
         }
